fix build in plugin terra basic caused by the httpInterceptorProviders
